Move OpenPoll utils to the OpenPollMsg module
